Functions of the fields are moved to a separate file
